// src/types/asset-info.ts
export type AssetInfo = {
    asset_string: string | null;
    code: string | undefined;
    home_domain: string | null;
    image: string | null;
    issuer: string;
    name: string;
    auth_required?: boolean;
    desc?: string;
    accounts_authorized?: number;
    anchor_asset: string;
    anchor_asset_type: string;
    auth_clawback_enabled: boolean;
    auth_immutable: boolean;
    auth_revocable: boolean;
    balances_authorized: string;
    claimable_balances_amount: string;
    conditions: string;
    is_asset_anchored: boolean;
    is_verified: boolean;
    liquidity_pools_amount: string;
    is_supply_locked: boolean;
    first_transaction: string;
};